⚡ Fix: 알람 세팅 엔티티 컬럼 및 자체 로그인 API 변경

+++ src/main/java/com/poppin/poppinserver/report/dto/report/response/ReportedReviewDetailDto.java
@@ -1,14 +1,13 @@
 package com.poppin.poppinserver.report.dto.report.response;
 
-import lombok.Builder;
-
 import java.util.List;
+import lombok.Builder;
 
 @Builder
 public record ReportedReviewDetailDto(
         Long reviewId,
         String reviewWriter,
-        Long reviewCnt,
+        Integer reviewCnt,
         String reviewContent,
         String reviewCreatedAt,
         boolean isCertificated,
