+++ src/main/java/com/poppin/poppinserver/user/dto/user/response/UserMypageDto.java
@@ -2,11 +2,12 @@ package com.poppin.poppinserver.user.dto.user.response;
 
 import lombok.Builder;
 
+// TODO: 리팩 후 삭제예정
 @Builder
 public record UserMypageDto(
         String nickname,
         String userImageUrl,
-        Long writtenReview,
-        Long visitedPopupCnt
+        Integer writtenReview,
+        Integer visitedPopupCnt
 ) {
 }
